requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.24.4
pandas==2.0.3
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2023.3.post1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.16.0
tzdata==2023.4
urllib3==1.25.10
Werkzeug==1.0.1